--- Assignment/Assignment/Assignment.csproj
@@ -1,1 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk" />
