--- Assignment/Assignment/PingProcess.cs
@@ -1,214 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Diagnostics;
-using System.Linq;
-using System.Text;
-using System.Threading;
-using System.Threading.Tasks;
-
-namespace Assignment;
-
-public record struct PingResult(int ExitCode, string? StdOutput);
-
-public class PingProcess
-{
-    private ProcessStartInfo StartInfo { get; } = new("ping");
-
-    public PingResult Run(string hostNameOrAddress)
-    {
-        StartInfo.Arguments = hostNameOrAddress;
-        StringBuilder? stringBuilder = null;
-        void updateStdOutput(string? line) =>
-            (stringBuilder??=new StringBuilder()).AppendLine(line);
-        Process process = RunProcessInternal(StartInfo, updateStdOutput, default, default);
-        return new PingResult( process.ExitCode, stringBuilder?.ToString());
-    }
-
-    public Task<PingResult> RunTaskAsync(string hostNameOrAddress)
-    {
-        return Task.Run(() => Run(hostNameOrAddress));
-    }
-
-    async public Task<PingResult> RunAsync(
-        string hostNameOrAddress, CancellationToken cancellationToken = default)
-    {
-        Task<PingResult> task = Task.Run (() => 
-            {
-                PingResult result = Run(hostNameOrAddress);
-                cancellationToken.ThrowIfCancellationRequested();
-                return result;       
-            }
-            ,cancellationToken);
-        return await task;
-    }
-
-    async public Task<PingResult> RunAsync(IEnumerable<string> hostNameOrAddresses, CancellationToken cancellationToken = default)
-    {
-        StringBuilder? stringBuilder = new StringBuilder();
-        var objectlock = new object();
-
-        try
-        {
-            var task = hostNameOrAddresses.Select(async hostOrAddress =>
-            {
-
-                cancellationToken.ThrowIfCancellationRequested();
-                PingResult result = await RunAsync(hostOrAddress, cancellationToken);
-
-                lock (objectlock)
-                {
-                    if (!string.IsNullOrEmpty(result.StdOutput))
-                    {
-                        stringBuilder.AppendLine(result.StdOutput.Trim());
-                    }
-                }
-                return result.ExitCode;
-
-            });
-
-            int[] exitCodes = await Task.WhenAll(task);
-
-            int combinedExitCode = exitCodes.Sum();
-
-            return new PingResult(combinedExitCode, stringBuilder.ToString());
-        }
-        catch (OperationCanceledException ex)
-        {
-            throw new AggregateException(new TaskCanceledException("Task Ended", ex));
-        }
-    }
-
-            
-        
-
-        /*ParallelQuery<Task<int>>? all = hostNameOrAddresses.AsParallel().Select(async item =>
-        {
-            Task<PingResult> task = null!;
-            // ...
-
-            await task.WaitAsync(default(CancellationToken));
-            return task.Result.ExitCode;
-        });
-
-        await Task.WhenAll(all);
-        int total = all.Aggregate(0, (total, item) => total + item.Result);
-        return new PingResult(total, stringBuilder?.ToString());
-    }*/
-
-    async public Task<PingResult> RunLongRunningAsync(
-        string hostNameOrAddress, CancellationToken cancellationToken = default)
-    {
-        Task task = null!;
-        await task;
-        throw new NotImplementedException();
-    }
-
-    private Process RunProcessInternal(
-        ProcessStartInfo startInfo,
-        Action<string?>? progressOutput,
-        Action<string?>? progressError,
-        CancellationToken token)
-    {
-        var process = new Process
-        {
-            StartInfo = UpdateProcessStartInfo(startInfo)
-        };
-        return RunProcessInternal(process, progressOutput, progressError, token);
-    }
-
-    private Process RunProcessInternal(
-        Process process,
-        Action<string?>? progressOutput,
-        Action<string?>? progressError,
-        CancellationToken token)
-    {
-        process.EnableRaisingEvents = true;
-        process.OutputDataReceived += OutputHandler;
-        process.ErrorDataReceived += ErrorHandler;
-
-        try
-        {
-            if (!process.Start())
-            {
-                return process;
-            }
-
-            token.Register(obj =>
-            {
-                if (obj is Process p && !p.HasExited)
-                {
-                    try
-                    {
-                        p.Kill();
-                    }
-                    catch (Win32Exception ex)
-                    {
-                        throw new InvalidOperationException($"Error cancelling process{Environment.NewLine}{ex}");
-                    }
-                }
-            }, process);
-
-
-            if (process.StartInfo.RedirectStandardOutput)
-            {
-                process.BeginOutputReadLine();
-            }
-            if (process.StartInfo.RedirectStandardError)
-            {
-                process.BeginErrorReadLine();
-            }
-
-            if (process.HasExited)
-            {
-                return process;
-            }
-            process.WaitForExit();
-        }
-        catch (Exception e)
-        {
-            throw new InvalidOperationException($"Error running '{process.StartInfo.FileName} {process.StartInfo.Arguments}'{Environment.NewLine}{e}");
-        }
-        finally
-        {
-            if (process.StartInfo.RedirectStandardError)
-            {
-                process.CancelErrorRead();
-            }
-            if (process.StartInfo.RedirectStandardOutput)
-            {
-                process.CancelOutputRead();
-            }
-            process.OutputDataReceived -= OutputHandler;
-            process.ErrorDataReceived -= ErrorHandler;
-
-            if (!process.HasExited)
-            {
-                process.Kill();
-            }
-
-        }
-        return process;
-
-        void OutputHandler(object s, DataReceivedEventArgs e)
-        {
-            progressOutput?.Invoke(e.Data);
-        }
-
-        void ErrorHandler(object s, DataReceivedEventArgs e)
-        {
-            progressError?.Invoke(e.Data);
-        }
-    }
-
-    private static ProcessStartInfo UpdateProcessStartInfo(ProcessStartInfo startInfo)
-    {
-        startInfo.CreateNoWindow = true;
-        startInfo.RedirectStandardError = true;
-        startInfo.RedirectStandardOutput = true;
-        startInfo.UseShellExecute = false;
-        startInfo.WindowStyle = ProcessWindowStyle.Hidden;
-
-        return startInfo;
-    }
-}
